+++ examples/seaice.py
@@ -22,7 +22,7 @@ cdef("frequency","monthly") ;
 cdef("project","EM")
 
 # Define your dataset (a number of facets take default values)
-sic=ds(experiment="NG89", variable="sic", period="198310", realm="I")
+sic=ds(simulation="NG89", variable="sic", period="198310", realm="I")
 
 # Display the basic filenames involved in the dataset (all filenames in one single string)
 # CliMAF will search them at the data location which is the most specific among all declared data locations 
@@ -40,7 +40,7 @@ os.system("ls -al "+my_file)
 os.system("type ncdump && ncdump -h "+my_file)
 
 # Stereopolar plot with explicit levels. See other plot arguments in on-line help : 'help(plot)'
-fig1=plot(sic,title="Sic 198310 2",proj="NH70",levels="0.1 10 30 60 80 90 95 97 98 99", contours=1, focues="ocean")
+fig1=plot(sic,title="Sic 198310 2",proj="NH70",levels="0.1 10 30 60 80 90 95 97 98 99", contours=1, focus="ocean")
 cshow(fig1)
 
 # Create a dictionnary in order to simplify the way to provide constant plot args :
@@ -52,13 +52,13 @@ figm=plot(masked_sic,title="mask above 0.99",**pa)
 cshow(figm)
           
 # Remap Sea Ice data to atmospheric grid. Use default option (remapbil)
-tas=ds(project="EM", experiment="GSAGNS1", variable="tas", period="1975", realm="L")
+tas=ds(project="EM", simulation="GSAGNS1", variable="tas", period="1975", realm="L")
 sic_on_tas_grid=regrid(sic,tas)
 fig2=plot(sic_on_tas_grid,title="regridded_sic",**pa)
 cshow(fig2)
 
 # Access daily data
-sicd=ds(experiment="NG89", variable="sic", period="198303-198304", realm="I", frequency='daily')
+sicd=ds(simulation="NG89", variable="sic", period="198303-198304", realm="I", frequency='daily')
 print sicd.baseFiles()